moved profile related components into profile folder
